forwardport of rev 3134, 3135 from 1.0 branch
git-svn-id: svn://svn.camaya.net/gloox/trunk@4136 bb60cc1b-8fe5-0310-9879-aca5d778a633

+++ src/connectioncompression.cpp
@@ -120,7 +120,7 @@ namespace gloox
     m_state = StateDisconnected;
   }
 
-  void ConnectionCompression::getStatistics( int& totalIn, int& totalOut )
+  void ConnectionCompression::getStatistics( long int& totalIn, long int& totalOut )
   {
     if( m_connection )
       m_connection->getStatistics( totalIn, totalOut );
