+++ src/lastactivityhandler.h
@@ -42,7 +42,7 @@ namespace gloox
        * on the JID, check the spec.)
        * @param status If the contact is offline, this is the last presence status message. May be empty.
        */
-      virtual void handleLastActivityResult( const JID& jid, int seconds, const std::string& status ) = 0;
+      virtual void handleLastActivityResult( const JID& jid, long seconds, const std::string& status ) = 0;
 
       /**
        * This function is called when an error is returned by the queried antity.
